// src/common/types/api.d.ts
export type PaginatedResponse<T> = {
  success?: boolean
  message?: string
  data: T[]
  meta: {
    current_page: number
    last_page: number
    per_page: number
    total: number
  }
  links?: {
    first?: string | null
    last?: string | null
    prev?: string | null
    next?: string | null
  }
}